Add files via upload
Built all code into the CleanCaseStudyData.R file.  Also split source code into modular parts like the assignment asked for on the second page in the Code section.  Designed so by running the Makefile.R file it calls all the other source files and runs them. One for downloading, one for cleaning, and one for merging the data.  I used the same technique and format that was used in the Week 8 section 4 video.